Prefix all generated fields with an underscore
This makes it possible to have field names that are keywords in rust.
The fields themselves have always been private, so this is not a breaking
change, with one exception: fields that are named "type" previously had
generated accessors of the form "{}_field_type" instead of the standard
"{}_type". This is no longer the case.

I've added a test for this specific case (field names), but I assume that
other cases (message names, enum names, enum variants) are still broken
when using rust keywords.

Fixes #105