Adds `SpecialFields` struct
The purpose is to reduce a little the automatically generated code.

The new struct implements PartialEq and is supposed to replace the *special* fields
- cached_size
- unknown_fields

On top of reducing the code (can derive PartialEq), it also hide the technical
implementation detail of cached_size's `Cell`.

Also adds a default implementation for `type_id` fn.