Move away from unsafe copy_nonoverlapping in write_raw_bytes
This commit removes the use of copy_nonoverlapping in favor of
copy_from_slice as per the TODO comment left in the source code.
The stable version of this writing is 1.14.0.

I am unsure how to run the performace testing code in src/perftest
and do not have a comparision to offer.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>