Treat paths as impl AsRef<Path>
APIs in terms of Path are more convenient for build scripts that
programmatically construct paths to input files or include-directories
using the std::path library. As an example of such code, Chromium OS
contains the following in one of its build scripts:

    let proto_root = match env::var("SYSROOT") {
        Ok(dir) => PathBuf::from(dir).join("usr/include/chromeos"),
        // Make this work when typing "cargo build" in platform2/vm_tools/crostini_client
        Err(_) => PathBuf::from("../../system_api"),
    };

    let concierge = proto_root.join("dbus/vm_concierge/service.proto");
    let cicerone = proto_root.join("dbus/vm_cicerone/service.proto");
    let seneschal = proto_root.join("dbus/seneschal/service.proto");
    let input_files = [concierge, cicerone, seneschal];

In order to pass those paths along to protoc_rust, they require an ugly
and panicky conversion from Path to &str.

    fn paths_to_strs(paths: &[PathBuf]) -> Vec<&str> {
        paths
            .iter()
            .map(|p| p.as_os_str().to_str().unwrap())
            .collect()
    }

    protoc_rust::Args {
        input: &paths_to_strs(&input_files),
        ...
    }

This commit adjusts the API of protoc_rust and other entry points to
accept AsRef<Path> as the first-class type of anything that represents a
filesystem path.

In order to preserve convenient support for the common case of hardcoded
string literal paths, we change code that used to use Args as a struct
literal into a builder instead. This aligns with the API of the `cc`
crate, and is similar to the std::process::Command API as well.

Before:

    protoc_rust::run(protoc_rust::Args {
        out_dir: "src/protos",
        input: &["protos/a.proto", "protos/b.proto"],
        includes: &["protos"],
        customize: Customize::default(),
    }).expect("protoc");

After:

    protoc_rust::Args::new()
        .out_dir("src/protos")
        .inputs(&["protos/a.proto", "protos/b.proto"]),
        .include("protos")
        .run()
        .expect("protoc");

The relevant builder method signatures are:

    impl Args {
        fn new() -> Self;
        fn out_dir(&mut self, out_dir: impl AsRef<Path>) -> &mut Self;
        fn include(&mut self, include: impl AsRef<Path>) -> &mut Self;
        fn includes(&mut self, includes: impl IntoIterator<Item = impl AsRef<Path>>) -> &mut Self;
        fn input(&mut self, input: impl AsRef<Path>) -> &mut Self;
        fn inputs(&mut self, inputs: impl IntoIterator<Item = impl AsRef<Path>>) -> &mut Self;
        fn customize(&mut self, customize: Customize) -> &mut Self;
        fn run(&self) -> Result<()>;
    }