implement Error for ParseErrorWithLoc
Fixes #422

This allows me to just use ? on such an error from a function that
returns failure::Error.

In the issue, I also mentioned that ParseErrorWithLoc is a private type.
(It's in a private mod and is never re-exported.) Looks like that
doesn't matter if I just want to use its Error/Display/Debug
implementations, so I didn't change it.