Use vector, not hashmap, to back UnknownFields
To match protoc, the iteration order of unknown fields must be stable
and match the order in which the unknown fields were encountered. The
C++ library uses a vector to back the structure, despite that making
many operations O(n)--it's unlikely to matter given how few fields are
in most protobuf messages.