Keep fields in a file descriptor proto in source code order
The fields in a file descriptor proto are emitted in source code order
by protoc. protoc-codegen-pure was reordering fields when normal fields
were interleaved with oneofs. Restructure the AST so there is one list
of both normal and oneof fields, entitled MessageEntry, so that the
order can be easily preserved.