Embed rustproto.proto and WKT protos in protobuf-codegen-pure
Embed the definitions of rustproto.proto and the well-known type protos
in the protobuf-codegen-pure crate. This ensures that the correct
versions of these proto files are available in every installation of
protobuf-codegen-pure. (Previously, users needed to acquire their own
copy of these protos and configure their include paths accordingly.)

Users can still override these embedded definitions by supplying a file
of the correct name, which will be preferred to the embedded
definitions.

This is a backport of aed791f7d.