Fix invalid aliasing and uninint memory in streams
Using remove_lifetime_mut to construct then hold on to both a Vec and a
mutable slice into it violates the existing Rust aliasing rules,
because both &mut and Unique<T> (which is used to implement Vec) assert
unique access to their contents. This problem is flagged by Miri running
cargo miri test. Though at the moment Stacked Borrows is experimental,
the aforementioned properties are already relied upon by rustc.

Getting access to the uninitialized region of a Vec as a &mut[T] is not
permitted by passing invalid indexes to get_unchecked_mut, or by any other
mechanism because slice::from_raw_parts documents that it requires
properly initialized values of T. Though any bit pattern is a valid u8,
and we never do a read, it is still invalid to construct the slice.