Advance pos_of_buffer_start when writing directly to a writer
If a message is bigger than the buffer, it will be written
directly to the writer, and there was a bug in that case where
pos_of_buffer_start was not advanced. This causes later
assertions to fail, such as the one in
Message::write_length_delimited_to, which asserts based on
total_bytes_written(), which is in turn based on
pos_of_buffer_start.