Make #[non_exhaustive] for oneofs configurable
Add a configuration option to disable annotating generated enum types
for proto `oneof`s with `#[non_exhaustive]`. The annotation makes sense
for self-contained crates, but it's convenient to be able to turn it
off.

One use case is for larger projects that consist of multiple crates that
are developed together. Being unable to exhaustively match on the
contents of `oneof`s not just in the crate that includes the generated
protobuf code is very useful.