Derive Hash for EnumOrUnknown
Since EnumOrUnknown does not have a protoc insertion point, additional
derives cannot be added via the api. At the same time, EnumOrUnknown are
simple values that are likely to be stored in a hash collection, such as
a map or set. This single word commit adds the derive for the Hash trait
to EnumOrUnknown.