chore: apply fixes to all the codebase with `cargo clippy --fix`
Checking the source code with Clippy is a good practice that helps maintaining the code in good shape and ensures an homogeneous style. Most of the automatic fixes in this commit are small cosmetic changes, but some of them improve legibility or even performance.

If this change is accepted I plan to keep applying more fixes for hints that Clippy reports but can't fix automatically.