Use destructuring instead of field point access
This ensures that if a new field is added in the future, this code will
be updated to handle it (one way or another), because not doing so will
lead to a compilation error. This will prevent issues like what happened
with oneof_nonexhaustive, where the field was added but its value was
not propagated.